Move testing helpers to internal directory
